feat: Add IDBBackendPoolService
This commit introduces the `IDBBackendPoolService` class which implements the `IBackendPoolService` interface using `indexedDB` backend. The class pools the database connections and provides the `resolve` and `release` methods to acquire and release a database connection for `backend` and `registry`. The `IDBBackendServiceFactory` is removed and `IDBBackendService` now takes care of opening and closing the DB connections. Removed unused `IDBService` and `close` method from `IDBBackendService`. The `RegistryService` now resolves `backend` for a given `workspace` using the new `IDBBackendPoolService`.