feat: refactor codecs, add PlanetScaleBackendManagerService
This commit makes changes to the `workspaceDataCodec` and `eventCodec` services by moving their definitions to their respective files. It also adds a new implementation of `IBackendManagerService` called `PlanetScaleBackendManagerService` which will be used to interact with a database service called PlanetScale. PlanetScaleBackendManagerService is equipped with new codec services to decode and encode data objects.